ㅇㅇㅇ

+++ src/main/java/com/example/winwin/mapper/board/CsMapper.java
@@ -26,11 +26,10 @@ public interface CsMapper {
     List<CsVo> selectScroll(CsVo csVo);
 //    메인페이지 갯수 구하기
     int selectTotal();
-
 //    조회수 증가
     public int upHit(Long csNumber);
-//    댓글수조회
-    public int commentCnt(Long csNumber);
+//    댓글수 증가
+    public Long commentCnt(Long csNumber);
 //   유저 프로필
     public CsProfileVo userLogin(Long userNumber);
 }
